Fix re-join bug. ...
When a node tries to re-join the cluster (because, say, it hard-crashed),
it is supposed to eject itself from the cluster first, because progress
can't be made on the cluster state without the participation of the
re-joining node, and the node can't finish the join and therefore can't
participate until the cluster state progresses to the point of the
join. This leads to deadlock if the re-joining peer hasn't been fully
ejected. This change completes the ejection process by acknowledging
the initial disassociate on behalf of the re-joining peer.